begins work on prehandler-router idea
