--- cell/eventListeners.js
@@ -1,27 +1,0 @@
-'use strict';
-
-const CellHandlers = require('./eventHandlers');
-
-
-exports.cellMousedownListener = (state, cell) => {
-
-    CellHandlers.handleCellMousedown(state, cell);
-};
-
-
-exports.cellMouseoverListener = (state, cell) => {
-
-    CellHandlers.handleDrag(state, cell);
-};
-
-
-exports.cellInputListener = (state, cell) => {
-
-    if (state.funcCellOutput[cell.id]) {
-        CellHandlers.handleFuncCellOutput(state, cell);
-    }
-    if (state.funcCellInput[cell.id]) {
-        CellHandlers.handleFuncCellInput(state, cell);
-    }
-    return;
-}
