+++ lib/cell/common.js
