+++ lib/cell/stateUpdate.js
