+++ lib/cell/styles.js
