+++ lib/eventListener.js
@@ -117,7 +117,13 @@ const windowKeydown = (e) => {
 };
 
 
-const windowMouseup = () => new Handler({ type: 'windowMouseup' });
+const windowMouseup = () => {
+
+    new Handler({ type: 'windowMouseup' }
+});
 
 
-const windowKeyup = () => new Handler({ type: 'windowKeyup' });
+const windowKeyup = () => {
+
+    new Handler({ type: 'windowKeyup' }
+});
