+++ lib/handlers/cellDrag.js
@@ -1,19 +1,11 @@
 'use strict';
 
 const { $updateCell } = require('../state');
-const CellCommon = require('./common');
-const CellStateUpdate = require('./stateUpdate');
+const CellCommon = require('../cell/common');
+const CellStateUpdate = require('../cell/stateUpdate');
 const WindowStateUpdate = require('../window/stateUpdate');
 
-
-const handleCellMousedown = (state, cell) => {
-
-    CellCommon.newSelectedCell(state, cell);
-    WindowStateUpdate.toggleMousedown(true)
-};
-
-
-const handleDrag = (state, cell) => {
+module.exports = (state, cell) => {
 
     const start = state.startCellRect;
     const { left, top, width, height } = CellCommon.getMultiCellDimensions(state, start, cell);
@@ -59,27 +51,3 @@ const handleDrag = (state, cell) => {
     });
     return;
 };
-
-
-const handleFuncCellInput = (state, cell) => {
-
-    state.funcCellInput[cell.id].forEach((inputCellId) => {
-
-        const cellToUpdate = state.allCells[inputCellId];
-        CellStateUpdate.updateFuncCellInputValue(state, cellToUpdate);
-    })
-};
-
-
-const handleFuncCellOutput = (state, cell) => {
-
-    CellStateUpdate.updateFuncCellOutputValue(state, cell);
-};
-
-
-module.exports = {
-    handleCellMousedown,
-    handleDrag,
-    handleFuncCellInput,
-    handleFuncCellOutput
-}
