+++ lib/handlers/enableCommandActive.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const { $setState } = require('../state');
+
+
+module.exports = (state) => {
+
+    $setState({ commandActive: true });
+
+    return;
+};
