+++ lib/handlers/funcCellOutput.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const CellStateUpdate = require('../cell/stateUpdate');
+
+module.exports = (state, cell) => {
+
+    CellStateUpdate.updateFuncCellOutputValue(state, cell);
+};
