+++ lib/handlers/keyup.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const { $setState } = require('../state');
+
+
+module.exports = () => {
+
+    $setState({ commandActive: false });
+
+    return;
+};
