+++ lib/handlers/mouseup.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const { $setState } = require('../state');
+
+module.exports = (state) => {
+
+    $setState({
+        mousedown: false,
+        colDrag: false,
+        rowDrag: false
+    });
+
+    return;
+};
