+++ lib/toolbar/common.js
