+++ lib/toolbar/eventListeners.js
