+++ lib/toolbar/styles.js
