+++ lib/window/common.js
