+++ lib/window/stateUpdate.js
